Transforms are faster
If you were wondering how much faster the ImprovedLocalToParentSystem is compared to Unity's when all the transforms are dirty, it is 4%. That's not much, but that is also worst-case scenario. I thought the performance would be identical, but either there's less change version checking or passing the ltw matrix by reference on the stack is a small win.

Now for the extreme transforms, those got some interesting results. I reworked the chunk lists to run in parallel, bringing the job count down from 50-some to 38 total. That made a big difference, because now the jobs are all parallel and run flush with each other. Without structural changes, they run twice as fast!

Now 10000 exposed transform characters are hitting 60FPS. Still plenty of improvement left though.