Hideous Hacks for Clean Installers
Wow is the hack for ICustomConversionBootstrap ugly. I am relying on the exact source code behavior of Entities 0.50 to get it right. That involves lots of juggling of event subscriptions to get useful entry points into the code. Plus Unity is really fighting me to try and enable every conversion system it can. But it seems to be working now.

As for the installers API, those came out clean! I even figured out how to make sceneBlackboardEntity user-resettable if scene management is not installed. This should solve a lot of problems users have with compatibility.

Next up I need to make the new bootstrap templates. And then I need to figure out what it takes to get NetCode working. And then it will be exploration of the Physics SmartBlobbers and migrating material instance property uploads into culling.