Dynamic Deform Buffer Indices
This is the third public test alpha.

100,000 compute deformed skinned dancers! They render. They actually render. That is well beyond what the Hybrid Renderer can handle.

The secret is that I am only allocating VRAM for what is actually on screen for any given frame. So while the number of entities grows, the number of entities on screen eventually reaches a limit. And that number is small enough that all the deformed vertices fit into 1 GB on the GPU.

This makes compute skinning with LODs completely viable, GPU-wise. However, at these extreme scales, I'm CPU-bound. However, I suspect that someone with a much newer CPU than me can probably hit 30 FPS with these 100,000 dancers.

Next up for weekend projects is the new binding system which handles sparse bone sets.