New Binding System (Big Effort)
Three weeks of nights and weekends, and it has finally come together. I'm sure there's still some dragons and pain points in this code, but the architecture is now where I want it for production use.

A big change was making renderers that only used a subset of the skeleton to work correctly without uploading the skeleton multiple times. Originally, my plan was to hash bindposes and do incremental updates. But after seeing how ugly that got, I started experimenting with moving the bindposes to the GPU. That paid off! Because it also unified the pipeline between compute skinning and what will be LBS vertex skinning. The compute shader handles setting up both. This does come with the cost of limiting the number of bones and transforms in a hierarchy. We'll see if that matters.

RenderDoc was a hero during debugging. Being able to inspect the buffers and find out I wasn't uploading vertices to the ranges I expected, as well as other stupid bugs, I don't think I would have gotten this working without it.

The optimized code path is working again. I tested imports pre-posed and it seems to handle that too. Of course the real highlight was throwing 100,000 dancers at it now that VRAM isn't an issue. 24 FPS! I bet someone with a better computer than me can make it hit 60 FPS.

Now it is just vertex LBS skinning, a custom allocator for animation buffers, and the animation clip plugin. Then beta while I write docs, and finally a release!

I'll announce the next alpha after vertex LBS skinning is in place.