Support Linear Blend Skinning Node
Fun fact: Struct IComponentData have values visible in the inspector only if their fields are public, even though the struct itself might be internal or private. I changed some things so that this is now visible.

I found a couple other minor bugs which I fixed, but the big highlight is that Linear Blend Skinning works. And it might just be a little faster. It is hard to say because everything is plenty fast. But if it is faster in some scenarios like this one, I'm glad I added it!

I decided to forget about the custom allocator and only support individual bone sampling for the initial release. Other than being a slightly slower approach, it doesn't really lose much in terms of flexibility. And it cuts out a lot of scope, so I can actually get this released. The other clip sampling modes can be added later without compromising API.

Next commit is going to be a little system cleanup. And then I will probably do some documentation for these weeknights.