Fix exposed bone bounds by changing culling technique
So it turns out that culling bones was dumb. Really dumb. All it takes is a stretched out triangle across the corner of the screen and bone culling will miss it.

The correct solution is to combine the bounds per skeleton, which I was originally trying to avoid due to the intense random accesses that would entail. However, I could still leverage culling indices to relay change filter statuses and have indices for writing to per-thread arrays.

This approach is better for higher shadow map counts, worse for high-entropy simulations, and a wash for 10000 dancers in this test project.

I'm still main-thread bound at the 1000-2000 dancer scale, so that will probably be priority for optimization in the future. For now, I'm just glad this fix isn't absolutely tanking performance like I feared.